+++ src/pool.rs
@@ -2,7 +2,7 @@ use std::fmt;
 use std::future::Future;
 use std::ops::{Deref, DerefMut};
 use std::pin::Pin;
-use std::sync::{Arc, Weak};
+use std::sync::Arc;
 use std::time::Instant;
 
 use crate::builder::Builder;
@@ -188,7 +188,6 @@ impl<M: Manager + Send> ManagedPool<M> {
 }
 
 pub type SharedManagedPool<M> = Arc<ManagedPool<M>>;
-pub type WeakSharedManagedPool<M> = Weak<ManagedPool<M>>;
 
 pub struct Pool<M: Manager + Send>(SharedManagedPool<M>);